requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bleach==6.1.0
cssmin==0.2.0
Django==4.2.10
django-appconf==1.0.6
django-compressor==4.4
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
django-taggit==5.0.1
Jinja2==3.1.3
jsmin==3.0.1
MarkupSafe==2.1.5
pillow==10.2.0
pytz==2024.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
webencodings==0.5.1
